+++ packages/angular-workspace/projects/angular/src/directives/text-value-accessor.ts
@@ -1,4 +1,4 @@
-import { Directive, ElementRef } from '@angular/core';
+import { Directive, ElementRef, forwardRef } from '@angular/core';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 
 import { ValueAccessor } from './value-accessor';
@@ -7,13 +7,13 @@ import { ValueAccessor } from './value-accessor';
   /* tslint:disable-next-line:directive-selector */
   selector: 'rux-input, rux-textarea, rux-slider, rux-radio-group, rux-select',
   host: {
-    '(ruxinput)': 'handleChangeEvent($event.target.value)',
-    '(ruxchange)': 'handleChangeEvent($event.target.value)'
+    '(ruxinput)': 'handleChangeEvent($event.target?.["value"])',
+    '(ruxchange)': 'handleChangeEvent($event.target?.["value"])'
   },
   providers: [
     {
       provide: NG_VALUE_ACCESSOR,
-      useExisting: TextValueAccessor,
+      useExisting: forwardRef(() => TextValueAccessor),
       multi: true
     }
   ]