+++ packages/framework-testing/astro-angular/src/app/app.module.ts
@@ -1,12 +1,22 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { AstroComponentsModule } from '@astrouxds/angular';
 
 import { AppComponent } from './app.component';
+import { StandaloneFormComponent } from './standalone-form.component';
+import { StandaloneAstroComponent } from './standalone-astro.component';
 
 @NgModule({
   declarations: [AppComponent],
-  imports: [BrowserModule, AstroComponentsModule],
+  imports: [
+    BrowserModule,
+    AstroComponentsModule,
+    ReactiveFormsModule,
+    FormsModule,
+    StandaloneFormComponent,
+    StandaloneAstroComponent,
+  ],
   providers: [],
   bootstrap: [AppComponent],
 })